in_tail: Add support for multiline json docker logs (docker_mode_json)
docker_mode_json works only with docker_mode.
It adds support for parsing docker logs file in json format,
if single json log line was splitted on several line.

How it works:
The docker_mode_json assumes that every line in log is valid json.
If it cannot parse log line, it parses this line together
with the next one, and next one until the parsing is succesful.